requirements.txt
google-genai
python-dotenv
chromadb
SpeechRecognition
gTTS
pyaudio
gpiozero
lgpio
pvporcupine
pvrecorder
pyserial
adafruit-circuitpython-dht
pytest
pyyaml
